apps/interaction-data-service/pyproject.toml
[project]
name = "interaction-data-service"
version = "0.1.0"
description = "Interaction data service skeleton"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.133.1",
    "uvicorn[standard]>=0.41.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.43",
    "psycopg[binary]>=3.3.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
